// src/enums/error-messages.ts
export enum ErrorMessages {
  SOMETHING_WRONG = 'Oops! Something went wrong. Try again later.',
  DATABASE_SAVING = 'Error while saving the message:',
  CHAT_DATA = 'Error while getting the chat data:',
  GET_LLM_NAME = 'Error while getting the LLM name:',
  GET_SYSTEM_MESSAGE = 'Error while getting the system message:',
  FALSY_LLM_NAME = 'Model cannot be null or undefined',
  UPDATE_MODEL = 'Error while updating the model:',
  DELETE_CHAT_HISTORY = 'Error while deleting the chat history:',
}